# projects/Fraud-Response-Concierge/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Fraud-Response Concierge",
    "description": "Synthetic incidents only. No real bank actions.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/v1/actions/{action_id}/approve": {
      "post": {
        "summary": "Approve",
        "operationId": "approve_v1_actions__action_id__approve_post",
        "parameters": [
          {
            "name": "action_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Action Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Approval"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Approval": {
        "properties": {
          "expected_case_version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Expected Case Version"
          },
          "action_payload_hash": {
            "type": "string",
            "title": "Action Payload Hash"
          },
          "approval_challenge_id": {
            "type": "string",
            "title": "Approval Challenge Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "expected_case_version",
          "action_payload_hash",
          "approval_challenge_id"
        ],
        "title": "Approval"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}